// apps/web/src/api/models/reviewState.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * MESA Law API
 * OpenAPI spec version: 0.1.0
 */

export type ReviewState = typeof ReviewState[keyof typeof ReviewState];


export const ReviewState = {
  PROPOSED: 'PROPOSED',
  APPROVED: 'APPROVED',
  CORRECTED: 'CORRECTED',
  REJECTED: 'REJECTED',
  PUBLISHING: 'PUBLISHING',
  PUBLISHED: 'PUBLISHED',
  PUBLICATION_FAILED: 'PUBLICATION_FAILED',
} as const;